requirements.txt
torch
transformers
diffusers
gradio
pillow
numpy
torchvision
accelerate
ftfy
